feat: added aiogram-dialogs

+++ tgbot/services/utils.py
@@ -10,13 +10,12 @@ logger = logging.getLogger(__name__)
 async def delete_prev_message(state: FSMContext):
     user_data = await state.get_data()
 
-    if "prev_bot_message" in user_data:
-        try:
-            prev_bot_message: Message = user_data["prev_bot_message"]
-            await prev_bot_message.delete()
-            keyboard_message: Message = user_data["keyboard_message"]
-            await keyboard_message.delete()
-        except TelegramAPIError as e:
-            logger.exception(e.message)
-        except KeyError:
-            pass
+    try:
+        prev_bot_message: Message | None = user_data.get("prev_bot_message")
+        await prev_bot_message.delete() if prev_bot_message is not None else ...
+        keyboard_message: Message | None = user_data.get("keyboard_message")
+        await keyboard_message.delete() if keyboard_message is not None else ...
+    except TelegramAPIError as e:
+        logger.exception(e.message)
+    except KeyError:
+        pass
